[core] move the RayletClientInterface to its own build target
