+++ BUILD.bazel
@@ -130,7 +130,7 @@ ray_cc_library(
     ),
     deps = [
         "//src/ray/common:asio",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
     ],
 )
