+++ src/ray/core_worker/BUILD.bazel
@@ -312,7 +312,7 @@ ray_cc_library(
     hdrs = ["experimental_mutable_object_provider.h"],
     deps = [
         ":experimental_mutable_object_manager",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "//src/ray/rpc:client_call",
     ],
 )
@@ -336,7 +336,7 @@ ray_cc_library(
         ":reference_count",
         ":task_manager",
         "//src/ray/common:id",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_pool",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/synchronization",
     ],
