+++ src/ray/core_worker/lib/java/io_ray_runtime_task_NativeTaskExecutor.cc
@@ -20,7 +20,7 @@
 #include "ray/common/id.h"
 #include "ray/core_worker/common.h"
 #include "ray/core_worker/core_worker.h"
-#include "ray/raylet_client/raylet_client.h"
+#include "ray/raylet_client/raylet_client_interface.h"
 
 #ifdef __cplusplus
 extern "C" {
