+++ src/ray/core_worker/task_submission/BUILD.bazel
@@ -93,7 +93,7 @@ ray_cc_library(
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker:task_manager",
         "//src/ray/gcs:gcs_pb_util",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "//src/ray/rpc:core_worker_client",
         "@com_google_absl//absl/base:core_headers",
     ],
