+++ src/ray/core_worker/task_submission/normal_task_submitter.h
@@ -31,8 +31,8 @@
 #include "ray/core_worker/store_provider/memory_store/memory_store.h"
 #include "ray/core_worker/task_manager.h"
 #include "ray/core_worker/task_submission/dependency_resolver.h"
-#include "ray/raylet_client/raylet_client.h"
-#include "ray/rpc/node_manager/raylet_client_pool.h"
+#include "ray/raylet_client/raylet_client_interface.h"
+#include "ray/raylet_client/raylet_client_pool.h"
 #include "ray/rpc/worker/core_worker_client.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
