+++ src/ray/core_worker/task_submission/tests/BUILD.bazel
@@ -70,7 +70,7 @@ ray_cc_test(
         "//src/ray/common:test_util",
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker/task_submission:normal_task_submitter",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "//src/ray/rpc:core_worker_client",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
