+++ src/ray/core_worker/task_submission/tests/normal_task_submitter_test.cc
@@ -30,7 +30,7 @@
 #include "ray/common/task/task_util.h"
 #include "ray/common/test_util.h"
 #include "ray/core_worker/store_provider/memory_store/memory_store.h"
-#include "ray/raylet_client/raylet_client.h"
+#include "ray/raylet_client/raylet_client_interface.h"
 #include "ray/rpc/worker/core_worker_client.h"
 
 namespace ray {
