+++ src/ray/core_worker/tests/BUILD.bazel
@@ -75,7 +75,7 @@ ray_cc_test(
         "//src/ray/core_worker:memory_store",
         "//src/ray/core_worker:object_recovery_manager",
         "//src/ray/object_manager:object_manager_common",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
