+++ src/ray/gcs/gcs_server/BUILD.bazel
@@ -80,7 +80,7 @@ ray_cc_library(
         "//src/ray/protobuf:autoscaler_cc_proto",
         "//src/ray/protobuf:gcs_service_cc_proto",
         "//src/ray/protobuf:ray_syncer_cc_proto",
-        "//src/ray/rpc:node_manager_client",
+        "//src/ray/raylet_client:raylet_client_pool",
         "//src/ray/stats:stats_metric",
         "//src/ray/util:event",
         "//src/ray/util:logging",
@@ -289,9 +289,8 @@ ray_cc_library(
         "//src/ray/common:task_common",
         "//src/ray/raylet/scheduling:cluster_resource_scheduler",
         "//src/ray/raylet/scheduling:scheduling_context",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "//src/ray/rpc:core_worker_client",
-        "//src/ray/rpc:node_manager_client",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
     ],
@@ -398,9 +397,8 @@ ray_cc_library(
         "//src/ray/common:task_common",
         "//src/ray/protobuf:gcs_service_cc_proto",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
         "//src/ray/rpc:core_worker_client",
-        "//src/ray/rpc:node_manager_client",
         "//src/ray/util:logging",
         "@com_google_absl//absl/container:flat_hash_map",
         "@com_google_absl//absl/container:flat_hash_set",
@@ -523,10 +521,10 @@ ray_cc_library(
         "//src/ray/pubsub:publisher",
         "//src/ray/raylet/scheduling:scheduler",
         "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_pool",
         "//src/ray/rpc:core_worker_client",
         "//src/ray/rpc:gcs_server",
         "//src/ray/rpc:metrics_agent_client",
-        "//src/ray/rpc:node_manager_client",
         "//src/ray/util:counter_map",
         "//src/ray/util:exponential_backoff",
         "//src/ray/util:network_util",
