+++ src/ray/gcs/gcs_server/gcs_server.h
@@ -39,10 +39,10 @@
 #include "ray/gcs/store_client/redis_store_client.h"
 #include "ray/raylet/scheduling/cluster_lease_manager.h"
 #include "ray/raylet/scheduling/cluster_resource_scheduler.h"
+#include "ray/raylet_client/raylet_client_pool.h"
 #include "ray/rpc/client_call.h"
 #include "ray/rpc/gcs/gcs_rpc_server.h"
 #include "ray/rpc/metrics_agent_client.h"
-#include "ray/rpc/node_manager/raylet_client_pool.h"
 #include "ray/rpc/worker/core_worker_client_pool.h"
 #include "ray/util/throttler.h"
