+++ src/ray/raylet/BUILD.bazel
@@ -236,7 +236,6 @@ ray_cc_library(
         "//src/ray/pubsub:subscriber",
         "//src/ray/raylet/scheduling:scheduler",
         "//src/ray/rpc:core_worker_client",
-        "//src/ray/rpc:node_manager_client",
         "//src/ray/rpc:node_manager_server",
         "//src/ray/stats:stats_lib",
         "//src/ray/util:cmd_line_utils",
@@ -286,6 +285,8 @@ ray_cc_binary(
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/object_manager:ownership_object_directory",
         "//src/ray/raylet/scheduling:cluster_lease_manager",
+        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_pool",
         "//src/ray/rpc:metrics_agent_client",
         "//src/ray/stats:stats_lib",
         "//src/ray/util:cmd_line_utils",
