+++ src/ray/raylet/node_manager.cc
@@ -46,7 +46,6 @@
 #include "ray/raylet/scheduling/cluster_lease_manager.h"
 #include "ray/raylet/worker_killing_policy.h"
 #include "ray/raylet/worker_pool.h"
-#include "ray/rpc/node_manager/node_manager_client.h"
 #include "ray/stats/metric_defs.h"
 #include "ray/util/cmd_line_utils.h"
 #include "ray/util/event.h"
