+++ src/ray/raylet_client/BUILD.bazel
@@ -1,20 +1,55 @@
 load("//bazel:ray.bzl", "ray_cc_library")
 
-exports_files([
-    "raylet_client.h",
-])
+ray_cc_library(
+    name = "raylet_client_interface",
+    hdrs = [
+        "raylet_client_interface.h",
+    ],
+    deps = [
+        "//src/ray/protobuf:autoscaler_cc_proto",
+        "//src/ray/protobuf:common_cc_proto",
+        "//src/ray/protobuf:node_manager_cc_proto",
+        "//src/ray/rpc:client_call",
+    ],
+)
+
+ray_cc_library(
+    name = "raylet_client_pool",
+    srcs = ["raylet_client_pool.cc"],
+    hdrs = [
+        "raylet_client_pool.h",
+    ],
+    deps = [
+        ":raylet_client_interface",
+        "//src/ray/gcs/gcs_client:gcs_client_lib",
+    ],
+)
+
+ray_cc_library(
+    name = "node_manager_client",
+    hdrs = [
+        "node_manager_client.h",
+    ],
+    deps = [
+        ":raylet_client_interface",
+        "//src/ray/common:id",
+        "//src/ray/gcs/gcs_client:gcs_client_lib",
+        "//src/ray/protobuf:node_manager_cc_grpc",
+        "//src/ray/rpc:client_call",
+        "//src/ray/rpc:grpc_client",
+        "//src/ray/util:network_util",
+    ],
+)
 
 ray_cc_library(
     name = "raylet_client_lib",
     srcs = ["raylet_client.cc"],
     hdrs = ["raylet_client.h"],
     deps = [
-        "//src/ray/common:id",
-        "//src/ray/common:status",
-        "//src/ray/common:task_common",
+        ":node_manager_client",
+        ":raylet_client_interface",
         "//src/ray/flatbuffers:node_manager_generated",
         "//src/ray/protobuf:common_cc_proto",
-        "//src/ray/rpc:node_manager_client",
         "//src/ray/util:logging",
     ],
 )
