+++ src/ray/raylet_client/tests/BUILD.bazel
@@ -8,7 +8,8 @@ ray_cc_test(
     deps = [
         "//src/fakes/ray/rpc/raylet:fake_raylet_client",
         "//src/ray/gcs/gcs_client:gcs_client_lib",
-        "//src/ray/rpc:node_manager_client",
+        "//src/ray/raylet_client:node_manager_client",
+        "//src/ray/raylet_client:raylet_client_pool",
         "@com_google_googletest//:gtest",
         "@com_google_googletest//:gtest_main",
     ],
