+++ src/ray/rpc/BUILD.bazel
@@ -124,33 +124,6 @@ ray_cc_library(
 )
 
 ray_cc_library(
-    name = "node_manager_client",
-    srcs = ["node_manager/raylet_client_pool.cc"],
-    hdrs = [
-        "node_manager/node_manager_client.h",
-        "node_manager/raylet_client_pool.h",
-    ] + [
-        # TODO(eoakes): these are needed due to a circular dependency:
-        # raylet_client_pool.cc -> raylet_client.h -> node_manager_client.h
-        "//src/ray/raylet_client:raylet_client.h",
-    ],
-    visibility = ["//visibility:public"],
-    deps = [
-        ":client_call",
-        ":grpc_client",
-        "//src/ray/common:id",
-        "//src/ray/gcs/gcs_client:gcs_client_lib",
-        "//src/ray/protobuf:node_manager_cc_grpc",
-        "//src/ray/util:network_util",
-    ] + [
-        # TODO(eoakes): these three come from raylet_client.h, remove after breaking the circular dependency.
-        "//src/ray/ipc:client_connection",
-        "//src/ray/common:ray_object",
-        "//src/ray/common:task_common",
-    ],
-)
-
-ray_cc_library(
     name = "node_manager_server",
     hdrs = [
         "node_manager/node_manager_server.h",
@@ -248,7 +221,8 @@ ray_cc_library(
         "//src/ray/gcs/gcs_client:gcs_client_lib",
         "//src/ray/protobuf:core_worker_cc_grpc",
         "//src/ray/pubsub:subscriber",
-        "//src/ray/raylet_client:raylet_client_lib",
+        "//src/ray/raylet_client:raylet_client_interface",
+        "//src/ray/raylet_client:raylet_client_pool",
         "//src/ray/util:logging",
         "//src/ray/util:network_util",
         "@com_github_grpc_grpc//:grpc++",
