+++ src/ray/rpc/worker/core_worker_client_pool.h
@@ -24,8 +24,8 @@
 #include "absl/synchronization/mutex.h"
 #include "ray/common/id.h"
 #include "ray/gcs/gcs_client/gcs_client.h"
-#include "ray/raylet_client/raylet_client.h"
-#include "ray/rpc/node_manager/raylet_client_pool.h"
+#include "ray/raylet_client/raylet_client_interface.h"
+#include "ray/raylet_client/raylet_client_pool.h"
 #include "ray/rpc/worker/core_worker_client.h"
 
 namespace ray {
